CP/DSP/PE/TEV: fix the timing units, the EFB copy engine and the TEV bindings (issue #378)
Star Wars Rogue Squadron III faults on an instruction page before its MMU
handler can map the page, so the CP's reader had to stop guessing at the host
scheduler's progress: it follows the emulated time base now, and the WPAR
timeout that the real Gekko does not have is gone.

The time base counts the Gekko clock (486 MHz), so a tick is one instruction and
the decrementer steps by one; the two-tick counter was a slow-down hack from the
interpreter days. The DSP runs at 81 MHz, one instruction per six time base
ticks, so its anchor is a plain tick count and its batch is measured in
instructions instead of the half ticks that assumed a 40.5 MHz time base.

The copy engine's texture copy turns a rectangle of the EFB into a tiled texture
in main memory (gfx-pe.md 5.7). It was missing entirely, which left every title
that renders through the EFB drawing from textures that stayed zeroed. The
rectangle is re-packed into the destination format, the tile rows follow the
destination stride, and the rectangle is read in screen coordinates. Reading the
EFB is a pixel engine operation and moved there from GFXCore, which keeps only
whether the frame loop owns the GL context.

A copy's clear belongs to the copy: it turns the quads the copy read into the
clear colour. A texture copy's runs right after it; a display copy's waits for
the frame begin, because the backend shows the EFB where the hardware would show
the XFB the copy wrote. A frame can present in several copies and every one of
them clears its own rectangle, so they are all kept - keeping only the last one
left the bootrom's picture on screen.

A RAS1_TREF word binds a pair of TEV stages, and its halves are consumed by the
enabled stages in order (upper first). Reading a fixed half per stage parity made
a draw that textures in both stages sample the wrong image, which is what turned
every string of Prince of Persia's memory card dialog into a white bar: the
dialog keeps its font in map 1 and the stage that draws the text sampled the
button-prompt atlas in map 0.

+++ src/dspcore.cpp
@@ -788,9 +788,9 @@ namespace DSP
 		dsp->Suspend();
 	}
 
-	// Called by the CPU thread every Flipper tick step. Waking up on every step would mean about a
-	// million scheduler wakeups per second, so the thread is woken once per `DspWakeTicks` and then
-	// drains a whole batch.
+	// Called by the CPU thread every Flipper tick step. Waking up on every step would mean about
+	// half a million scheduler wakeups per second, so the thread is woken once per `DspWakeTicks`
+	// and then drains a whole batch.
 	void DspCore::TickSync(int64_t ticks)
 	{
 		if (ticks < wakeTick)
@@ -833,9 +833,9 @@ namespace DSP
 		// wakeup's worth. The anchor is taken again at the end exactly as the one-instruction-at-a-
 		// time version did: the DSP is limited by the host, not by the emulated clock, and the
 		// backlog is dropped rather than accumulated.
-		uint32_t budget = (uint32_t)(DspWakeTicks / (int64_t)GekkoTicksPerDspInstruction);
+		int64_t budget = DspWakeTicks / GekkoTicksPerDspInstruction;
 
-		while (budget > 0 && ticks >= (dsp->savedGekkoTicks + GekkoTicksPerDspInstruction))
+		while (budget > 0 && (int64_t)ticks >= dsp->savedGekkoTicks + GekkoTicksPerDspInstruction)
 		{
 			// Test breakpoints and canaries
 			if (dsp->IsRunning())
@@ -867,12 +867,12 @@ namespace DSP
 			Gekko::stats.dspInstrs += retired;
 			dsp->savedGekkoTicks += (int64_t)retired * GekkoTicksPerDspInstruction;
 
-			budget = (retired >= budget) ? 0 : (budget - retired);
+			budget = ((int64_t)retired >= budget) ? 0 : (budget - (int64_t)retired);
 		}
 
-		if (dsp->savedGekkoTicks <= ticks)
+		if (dsp->savedGekkoTicks <= (int64_t)ticks)
 		{
-			dsp->savedGekkoTicks = ticks;
+			dsp->savedGekkoTicks = (int64_t)ticks;
 		}
 	}
